requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart